+++ src/main/scala/Rt/Parser/Query.scala
@@ -8,7 +8,7 @@ object Query {
   val ticketsEmptyRe = """No matching results.""".r
   def parseQueryResponse( responseStr: String ) = {
     parseResponse( responseStr ).flatMap{
-      case ticketsEmptyRe()::ls => Nil.point[Parser]
+      case ticketsEmptyRe()::ls => List[Rt.QueryResult]().point[Parser]
       case lines => Field.parseFields( lines ).map(
         _.map( f => Rt.QueryResult( f.name.toInt , f.value ) )
       )