+++ src/main/scala/Rt/Parser/Read.scala
@@ -9,20 +9,20 @@ object Read {
   // Should turn this into a typeclass. It is pretty clunky.
 
   def readInt(s:String): String \/ Int = {
-    \/.fromTryCatch( s.toInt ).leftMap( _ => s"$s is not an int" )
+    \/.fromTryCatchNonFatal( s.toInt ).leftMap( _ => s"$s is not an int" )
   }
   def readLong(s:String): String \/ Long = {
-    \/.fromTryCatch( s.toLong ).leftMap( _ => s"$s is not an int" )
+    \/.fromTryCatchNonFatal( s.toLong ).leftMap( _ => s"$s is not an int" )
   }
   def readDouble(s:String): String \/ Double = {
-    \/.fromTryCatch( s.toDouble ).leftMap( _ => s"$s is not an int" )
+    \/.fromTryCatchNonFatal( s.toDouble ).leftMap( _ => s"$s is not an int" )
   }
   def readList(s:String): List[String] = {
     s.split(",").toList.map( _.trim )
   }
   def readDateTime(format: DateTimeFormatter, tz: DateTimeZone ) = {
     val tzFormat = format.withZone(tz)
-    def read(s:String) = \/.fromTryCatch(
+    def read(s:String) = \/.fromTryCatchNonFatal(
       tzFormat.parseDateTime(s)
     ).leftMap( t => s"$s is not a datetime. Err: ${t.getMessage}" )
     read _