+++ src/test/scala/Rt/Parser/Query.scala
@@ -9,7 +9,7 @@ object QueryParserSpec extends mutable.Specification {
     "Return Credentials required for 401" in {
       Query.parseQueryResponse(
         "RT/4.0.14 401 Credentials required"
-      ).run must beEqualTo(-\/(CredentialsRequired))
+      ) must beEqualTo(-\/(CredentialsRequired))
     }
     "Return empty list for no Results" in {
       Query.parseQueryResponse(
@@ -17,7 +17,7 @@ object QueryParserSpec extends mutable.Specification {
         |
         |No matching results.
         |""".stripMargin
-      ).run must beEqualTo(\/-(Nil))
+      ) must beEqualTo(\/-(Nil))
     }
     "Return list of query results" in {
       Query.parseQueryResponse(
@@ -26,7 +26,7 @@ object QueryParserSpec extends mutable.Specification {
         |948619: TADSL - Query regarding SQ results and possible WG release update - 0755730637
         |956459: FW: No network service 8GB on a 12GB Value Cap
         |""".stripMargin
-      ).run must beEqualTo(\/-(List(
+      ) must beEqualTo(\/-(List(
         QueryResult(948619,"TADSL - Query regarding SQ results and possible WG release update - 0755730637"),
         QueryResult(956459,"FW: No network service 8GB on a 12GB Value Cap")
       )))
@@ -39,7 +39,7 @@ object QueryParserSpec extends mutable.Specification {
               "" ::
               (1 to 10000).map( x => s"$x: Ticket number $x" ).toList
           ).mkString("\n")
-        ).map( _.length ).run must_==(\/-(10000))
+        ).map( _.length ) must_==(\/-(10000))
       } catch {
         case e:Throwable => e.printStackTrace() ; throw e
       }